+++ content-elements/form/form-container/form.js
@@ -10,13 +10,13 @@ Alpine.data('formElement', () => ({
     this.root = this.$root;
 
     if (this.root.classList.contains('bsi-form-label-floating')) {
-      for(const floatingElement of this.form.getElementsByClassName('bsi-label-floating-element')) {
+      for (const floatingElement of this.form.getElementsByClassName('bsi-label-floating-element')) {
         this._initFloatingLabels(floatingElement);
       }
     }
 
     if (this.root.classList.contains('bsi-form-info-as-tooltip')
-        && ['bsi-form-label-top', 'bsi-form-label-left'].some(labelClass => this.root.classList.contains(labelClass))) {
+      && ['bsi-form-label-top', 'bsi-form-label-left'].some(labelClass => this.root.classList.contains(labelClass))) {
       this.form.querySelectorAll('.bsi-form-element').forEach((formElement) => {
         let infoTextField = formElement.querySelector('.form-text');
         let tooltipIcon = formElement.querySelector('i');
@@ -35,6 +35,8 @@ Alpine.data('formElement', () => ({
       e.preventDefault();
       e.stopPropagation();
       this._validateRadioInput();
+      this._setAriaInvalid();
+      this._formValidationSummary();
     }
     this.form.classList.add('was-validated');
   },
@@ -47,7 +49,7 @@ Alpine.data('formElement', () => ({
       input = floatingElement.querySelector('.form-select');
     } else {
       if (input.placeholder.length === 0) {
-        if (input.type === "date") {
+        if (input.type === 'date') {
           label.style.opacity = 1;
           label.style.transform = 'initial';
         } else {
@@ -56,8 +58,8 @@ Alpine.data('formElement', () => ({
       }
     }
 
-    if (!(input.type === "range")) {
-      let floatingDiv = document.createElement("div");
+    if (!(input.type === 'range')) {
+      let floatingDiv = document.createElement('div');
       floatingDiv.classList.add('form-floating');
       floatingElement.insertBefore(floatingDiv, labelAndInfo.nextSibling);
       floatingDiv.append(input, label);
@@ -67,19 +69,27 @@ Alpine.data('formElement', () => ({
   _validateRadioInput() {
     let radioElements = this.form.getElementsByClassName('bsi-form-radio-element');
     for (const radioElement of radioElements) {
-      let radioValid = false;
-      let radioInputs = radioElement.getElementsByClassName('form-check-input');
-      for (const radioInput of radioInputs) {
-        if (radioInput.checked || !radioInput.hasAttribute('required')) {
-          radioValid = true;
-          break;
-        }
-      }
-      if (!radioValid) {
-        radioElement.getElementsByClassName('invalid-feedback')[0].style.display = "block";
-      } else {
-        radioElement.getElementsByClassName('invalid-feedback')[0].style.display = "none";
-      }
+      let radioInputs = Array.from(radioElement.querySelectorAll('.form-check-input'));
+      let radioValid = radioInputs.some(radio => radio.checked || !radio.hasAttribute('required'));
+      radioElement.getElementsByClassName('invalid-feedback')[0].style.display = radioValid ? 'none' : 'block';
+      radioElement.setAttribute('aria-invalid', !radioValid);
     }
   },
-}))
+
+  _setAriaInvalid() {
+    var elements = this.form.querySelectorAll('.bsi-form-element:not(.bsi-form-radio-element)');
+    elements.forEach(element => {
+      let input = element.querySelector('input, select, textarea');
+      input.setAttribute('aria-invalid', !input.checkValidity());
+    });
+  },
+
+  _formValidationSummary() {
+    if (this.root.classList.contains('bsi-form-show-valdiation-summary')) {
+      let invalidElements = Array.from(this.form.querySelectorAll('.bsi-form-element:has([aria-invalid=true]) .form-label-and-info-text label'));
+      let valdiationMessageHTML = invalidElements.map(label => `<li>${label.innerText}</li>`).join('');
+      let messageContainerList = this.form.querySelector('.form-validation-summary ul');
+      messageContainerList.innerHTML = valdiationMessageHTML;
+    }
+  },
+}));